watermark.go: aipr
