+++ go.mod
@@ -1,12 +1,10 @@
 module main
 
-go 1.21.3
+go 1.20
+
+require golang.org/x/image v0.16.0
 
 require (
-	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
-	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646
-	golang.org/x/image v0.16.0
-
+	github.com/fogleman/gg v1.3.0
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
 )
-
-require github.com/fogleman/gg v1.3.0 // indirect
